Change Milli_Virtual_Loss to Micro_Virtual_Loss

+++ engine/src/agents/config/searchsettings.h
@@ -54,7 +54,7 @@ struct SearchSettings
     float qValueWeight;
     // describes how much better the highest Q-Value has to be to replace the candidate move with the highest visit count
     float qVetoDelta;
-    float virtualLoss;
+    double virtualLoss;
     bool verbose;
     uint_fast8_t epsilonChecksCounter;
 //    bool enhanceCaptures;   currently not support
